pyproject.toml
[tool.poetry]
name = "gato-naranja_generate_diff"
version = "0.1.6"
description = "Second project on Hexlet"
authors = ["Alex Parshikov <alex.n.parshikov@gmail.com>"]
packages = [
  { include = "generate_diff" },
]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
pytest = "^6.1.1"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = "generate_diff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
